[MIRROR] Player Panel: Redux (#252)
* Player Panel: Redux (#8014)

* Player panel initial

* Stuff

* thing with no map

* Re add map

* Selection improve

* Final additions

* Fix

* Add a cooldown

* Fixes and optimizations

* Case insensitive

* Fix

* Improvements and performance

* Oops

* Fix

* Fix an error

* Bruh fix

* Tweak telemetry alert levels

* Optimize

* Remove View

* Swap PM and FLW

* Fixes

* Fix

* Bleh

* Make tooltips use one popper, fixing mount lag (#61783)

In the last episode, I fixed update lag. This is what made the R&D console laggy to scroll through.

In this episode, we fix mount lag, which is what makes the R&D console laggy to open and search through, since it has to constantly make the new tooltips.

This is done by using a trick tippy.js uses, where they create one popper instance for multiple tooltips, then just move it around. I decided to emulate this by moving the popper stuff to Tooltip itself. The Popper component still exists as some non-tooltip consumers for it exist, such as preferences menu clothing selection. This is done through the ugliest code of my fucking life, but in a way that's totally transparent to consumers.

* Minor improvements

* Fix lint error with null

* Add autofocus input and fix dragging borders being wrong

* Add telemetry locator

* Fix some line breaks

* Remove old player panel

Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>

* Player Panel: Redux

Co-authored-by: itsmeow <itsmeow@itsmeow.dev>
Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>